phase-27-issues-cleanup-l: resolve L1-L15 cleanup items
Pure cleanup, no behavior change:

- L1: Remove duplicate if (rc != VFS_OK) check in fuse_vfs_readdir
- L2: Remove redundant (void)mode/flags block and nested #ifdef __APPLE__
- L3: Collapse duplicate #pragma GCC diagnostic push with single
  comment explaining the macFUSE callback-signature incompatibility
- L4: Add a comment explaining why the crc32c table init runs on
  x86_64 even though the HW path is taken unconditionally
- L5: Remove cache_dump_dirty_by_priority no-op stub (declaration
  in storage.h and definition in page_cache.c)
- L6: Remove mirror_metrics_dump/mirror_metrics_pump undeclared
  declarations from storage.h
- L7: Update stale hash_map.c top-of-file comment to match the
  actual HASH_MAP_DEFAULT_SCALE=16 / GRANULARITY=9 defaults
- L8: Remove duplicate SB_OFF_SEGMENT_SIZE in tree.h (it shadows
  HDR_OFF_SEGMENT_SIZE in storage.h; the superblock is page 1, not
  page 0). Added a comment pointing to the canonical definition
- L9: SKIPPED — the dirchain_list code was refactored in Phase 26
  (now uses vfs_chain_walk for the read-rule); the stale issue
  refers to code that no longer exists
- L10: Document why fuse_vfs.c keeps its own vfs_error_to_str
  mirror instead of using the public vfs_error_string (include
  cycle avoidance)
- L11: SKIPPED — the epoch.c comment style issue refers to an
  editing artifact that no longer exists in the current code
- L12: Remove the garbled/truncated first 'DirNode (32 bytes, 16
  used, 16 reserved)' comment header in nodes.h; the second one
  (the real one) is kept
- L13: Fix 'DententryCache' typo in gc.h vfs_gc doc comment
  (replaced with 'slot cache')
- L14: SKIPPED — tcache perf note, the code comment already
  acknowledges the behavior
- L15: Add (size_t) cast to memset in page_array.c for consistency
  with the malloc just above

+++ src/page_array.c
@@ -10,7 +10,7 @@ int segment_array_build(Pool* pool, int64_t fc_pageRootPtr,
     if (!arr->vptr_array) return VFS_ERR_NOMEM;
 
     /* Fill with VFS_VPTR_NULL (0) — unallocated pages map to NULL */
-    memset(arr->vptr_array, 0, segment_size * sizeof(int64_t));
+    memset(arr->vptr_array, 0, (size_t)segment_size * sizeof(int64_t));
     arr->seg_size = segment_size;
 
     /* Walk the sparse PageNode chain, placing each node at its page_index */
